requirements.txt
fastcrc~=0.2.1
python-magic~=0.4.27
